+++ impl-24.sln
@@ -1,0 +1,33 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.002.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "temas", "temas", "{F7B61069-83AC-4785-A434-6FE7840DD5B7}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "delegacion", "delegacion", "{F687FE14-A0C5-4924-9240-DB01441476E5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "C#", "temas\delegacion\C#\C#.csproj", "{1328CB84-CE11-495C-8D8D-736F879FE2D1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1328CB84-CE11-495C-8D8D-736F879FE2D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1328CB84-CE11-495C-8D8D-736F879FE2D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1328CB84-CE11-495C-8D8D-736F879FE2D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1328CB84-CE11-495C-8D8D-736F879FE2D1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F687FE14-A0C5-4924-9240-DB01441476E5} = {F7B61069-83AC-4785-A434-6FE7840DD5B7}
+		{1328CB84-CE11-495C-8D8D-736F879FE2D1} = {F687FE14-A0C5-4924-9240-DB01441476E5}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {971D9969-73BA-49C7-BE1D-9E03BEA66D3F}
+	EndGlobalSection
+EndGlobal
