+++ temas/delegacion/Csharp/bin/Debug/net8.0/Csharp.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v8.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v8.0": {
+      "Csharp/1.0.0": {
+        "runtime": {
+          "Csharp.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "Csharp/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
